Adding a new interface file for RegEx_Get_CopyBook functio.
This new file adds 2 buttons "ok" and "cancel" as well as adding a menu to select files from a windows open prompt. The interface will also use registry to open the prompt in the last location used.

A recent files item allows users to select recently used files that remain in memory while the program is running.