Changed to use 2 entryfields
This allows minimal confustion as to what a valid parameter is. The main
window is not a proper GUI rather than an inputbox. The window will also
stay open even after a successful launch of the program to allow for
quicker opening of additional windows if needed.

With the two entry fields tool tips are used to explain what is to be
entered to allow for less window space needed.

Menu options may be available in later editions to allow users to select
y/n for setting default,
y/n for opening debug sessions,
y/n for initial debug 'go' click,
y/n for allowing login function,
y/n for use shorthand function
etc.

"shorthand" function would be a cycle of attempts based on only the first
numeric portion of the server name being entered.
ex:
	user enters 630
	program assumes "dds630vmub" and tries 01, then 02, then 03, etc
	until a successful session is opened