Adding a new file for predefined copy.
This script is very simple and will copy one file to another at a single
keystroke ence the window has been set with the source and destination the
user desires. The program includes a message box when first run to change
the title of the window to help users recognize a set of paths when using
multiple instances of the program.