Changed the catches to not stop the script
This will allow the greatest possible automation and the least required
user interaction. Known invalid vin messages and existing unit scenarios
are handled by deleting the data that caused the error and saving the
incomplete record, then recalling and deleting it. Issue causing records
are logged in a separate csv file (AgentsOfChaos.csv). The error for invalid
types requires user intervention because to recover the Advantage UI from
this error the program must be exited and the session reopened, as Advantage
locks the Unit # field.